chore: rename example app package and update documentation
- Renamed the example application package from `org.calypsonet.keyple.plugin.bluebird.example` to `org.calypsonet.keyple.example.plugin.bluebird` for better consistency.
- Updated `AndroidManifest.xml`, `gradle.properties`, and relevant Kotlin files to reflect the package name change.
- Renamed `DemoApplication` to `ExampleApplication`.
- Simplified `MainActivity` by removing redundant UI messages and observer cleanup logic.
- Expanded `README.md` with detailed sections on features, CNA library dependencies for storage card support, and MIFARE Classic key management.
- Updated `strings.xml` to rename the app to "Keyple Bluebird Example App".
