+++ example-app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
         android:name="com.bluebird.permission.SAM_DEVICE_ACCESS"/>
 
     <application
-        android:name=".DemoApplication"
+        android:name="org.calypsonet.keyple.example.plugin.bluebird.ExampleApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -14,7 +14,7 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name=".MainActivity"
+            android:name="org.calypsonet.keyple.example.plugin.bluebird.MainActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:clearTaskOnLaunch="true">
