+++ example-app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Keyple Bluebird Demo App</string>
+    <string name="app_name">Keyple Bluebird Example App</string>
 </resources>
