+++ gradle.properties
@@ -13,8 +13,8 @@ org.gradle.jvmargs = "-Dfile.encoding=UTF-8"
 
 # Android Configuration (custom properties)
 androidLibNamespace = org.calypsonet.keyple.plugin.bluebird
-androidAppNamespace = org.calypsonet.keyple.plugin.bluebird.example
-androidAppId = org.calypsonet.keyple.plugin.bluebird.example
+androidAppNamespace = org.calypsonet.keyple.example.plugin.bluebird
+androidAppId = org.calypsonet.keyple.example.plugin.bluebird
 androidAppVersionCode = 1
 androidAppVersionName = 1.0
 androidCompileSdk = 35
